fix: cleanup command deleting open branches (#193)
* fix: cleanup command deleting open branches

* feat: delete branches from closed pull requests

* feat: add only merged flag

+++ src/main/kotlin/commands/Cleanup.kt
@@ -1,10 +1,17 @@
 package commands
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.parameters.options.flag
+import com.github.ajalt.clikt.parameters.options.option
 import data.GitHubImpl
+import org.kohsuke.github.GHIssueState
 
 class Cleanup : CliktCommand(name = "cleanup") {
+    private val onlyMerged: Boolean by option().flag(default = false)
+
     override fun run() {
+        val mergeState = if (onlyMerged) "merged" else "merged or closed"
+        info("Deleting branches with a $mergeState pull request to ${GitHubImpl.wingetPkgsFullName} from them")
         val wingetPkgsFork = GitHubImpl.getWingetPkgsFork(currentContext.terminal)
         var branchesDeleted = 0
 
@@ -12,26 +19,36 @@ class Cleanup : CliktCommand(name = "cleanup") {
             .filterNot { it.name == GitHubImpl.microsoftWinGetPkgs.defaultBranch }
             .forEach { branch ->
                 runCatching {
-                    val commit = GitHubImpl.microsoftWinGetPkgs
-                        .queryCommits()
-                        .from(branch.shA1)
+                    val pullRequest = GitHubImpl.microsoftWinGetPkgs
+                        .queryPullRequests()
+                        .head("${wingetPkgsFork.ownerName}:${branch.name}")
+                        .base(GitHubImpl.microsoftWinGetPkgs.defaultBranch)
+                        .state(GHIssueState.CLOSED)
                         .list()
                         .withPageSize(1)
                         .first()
 
-                    commit.let {
-                        val branchName = branch.name
-                        warning("Deleting $branchName because it was merged in ${it.htmlUrl}")
-                        wingetPkgsFork.getRef("heads/$branchName").delete()
-                        branchesDeleted++
+                    pullRequest?.let {
+                        if (!onlyMerged || it.isMerged) {
+                            val branchName = branch.name
+                            val action = if (it.isMerged) "merged" else "closed"
+                            val url = if (it.isMerged) {
+                                "${GitHubImpl.microsoftWinGetPkgs.htmlUrl}/commit/${it.mergeCommitSha}"
+                            } else {
+                                it.htmlUrl
+                            }
+                            warning("Deleting $branchName because its pull request was $action in $url")
+                            wingetPkgsFork.getRef("heads/$branchName").delete()
+                            branchesDeleted++
+                        }
                     }
                 }
             }
 
-        if (branchesDeleted == 0) {
-            echo("No branches were found that could be deleted")
-        } else {
+        if (branchesDeleted > 0) {
             success("$branchesDeleted branches were deleted")
+        } else {
+            echo("No branches were found that could be deleted")
         }
     }
 }
